Crate http

Low level HTTP toolkit library

Feature "ssl" enable secured connection support with openssl. This feature is enabled by default but can be disabled

Modules

client
Module containing HTTP client implementations

constants
Defines some usefull constants

messages
HTTP messages definitions

methods
Methods definitions

streams
Http I/O streams definitions

Macros

debug!
Print debug message only if built in debug mode

map!
Quickly build a std::collections::HashMap

option!
Unwrap an Option<T> or return an Err if it's None

smap!
Build a std::collections::HashMap<String, String>

str!
Return a String if conversion is possible by calling to_string() on the provided expression

try_all!
Wrap all the given expressions in a try!

Enums

Protocol
Protocol implementations

Functions

open
Create a new client. The implementation which may differ according to the protocol is hidden behind a boxed Http trait