Struct http::buffer::BufferedStream

pub struct BufferedStream<T> {
    pub wrapped: T,
    pub read_buffer: Vec<u8>,
    pub read_pos: uint,
    pub read_max: uint,
    pub write_buffer: Vec<u8>,
    pub write_len: uint,
    pub writing_chunked_body: bool,
}

Fields

wrapped

read_buffer

read_pos

read_max

write_buffer

write_len

writing_chunked_body

Methods

impl<T: Stream> BufferedStream<T>

fn new(stream: T) -> BufferedStream<T>

impl<T: Reader> BufferedStream<T>

fn poke_byte(&mut self, byte: u8)

Poke a single byte back so it will be read next. For this to make sense, you must have just read that byte. If self.pos is 0 and self.max is not 0 (i.e. if the buffer is just filled Very great caution must be used in calling this as it will fail if self.pos is 0.

fn read_byte(&mut self) -> IoResult<u8>

Slightly faster implementation of read_byte than that which is provided by ReaderUtil (which just uses read())

impl<T: Writer> BufferedStream<T>

fn finish_response(&mut self) -> IoResult<()>

Finish off writing a response: this flushes the writer and in case of chunked Transfer-Encoding writes the ending zero-length chunk to indicate completion.

At the time of calling this, headers MUST have been written, including the ending CRLF, or else an invalid HTTP response may be written.