Module http::client::request

Things for the construction and sending of HTTP requests.

If you want to make a request, RequestWriter::new is where you start, and RequestWriter.read_response is where you will send the request and read the response.

extern crate http;
extern crate url;

use http::client::RequestWriter;
use http::method::Get;
use url::Url;

fn main() {
    let url = Url::parse("http://example.com/").unwrap();
    let request: RequestWriter = match RequestWriter::new(Get, url) {
        Ok(request) => request,
        Err(error) => fail!(":-( {}", error),
    };

    let mut response = match request.read_response() {
        Ok(response) => response,
        Err((_request, error)) => fail!(":-( {}", error),
    };
    // Now you have a `ResponseReader`; see http::client::response for docs on that.
}

If you wish to send a request body (e.g. POST requests), I'm sorry to have to tell you that there is not good support for this yet. However, it can be done; here is an example:

let data = b"var1=val1&var2=val2";
let mut request: RequestWriter = match RequestWriter::new(Get, url) {
    Ok(request) => request,
    Err(error) => fail!(":-( {}", error),
};

request.headers.content_length = Some(data.len());
request.write(data);
let response = match request.read_response() {
    Ok(response) => response,
    Err((_request, error)) => fail!(":-( {}", error),
};

Structs

RequestWriter