Struct http::client::request::RequestWriter

pub struct RequestWriter<S = NetworkStream> {
    pub remote_addr: Option<SocketAddr>,
    pub headers: Box<HeaderCollection>,
    pub method: Method,
    pub url: Url,
    // some fields omitted
}

Fields

remote_addr

The originating IP address of the request.

headers

The host name and IP address that the request was sent to; this must always be specified for HTTP/1.1 requests (or the request will be rejected), but for HTTP/1.0 requests the Host header was not defined, and so this field will probably be None in such cases. The headers sent with the request.

method

The HTTP method for the request.

url

The URL being requested.

Methods

impl<S: Reader + Writer = NetworkStream> RequestWriter<S>

Low-level HTTP request writing support

Moderately hacky, and due to current limitations in the TcpStream arrangement reading cannot take place until writing is completed.

At present, this only supports making one request per connection.

fn new(method: Method, url: Url) -> IoResult<RequestWriter<S>>

Create a RequestWriter writing to the specified location

fn new_request(method: Method, url: Url, use_ssl: bool, auto_detect_ssl: bool) -> IoResult<RequestWriter<S>>

impl<S: Connecter + Reader + Writer = NetworkStream> RequestWriter<S>

fn try_connect(&mut self) -> IoResult<()>

Connect to the remote host if not already connected.

fn connect(&mut self) -> IoResult<()>

Connect to the remote host; fails if already connected. Returns true upon success and false upon failure (also use conditions).

fn try_write_headers(&mut self) -> IoResult<()>

Write the Request-Line and headers of the response, if we have not already done so.

fn write_headers(&mut self) -> IoResult<()>

Write the Status-Line and headers of the response, in preparation for writing the body.

If the headers have already been written, this will fail. See also try_write_headers.

fn read_response(self) -> Result<ResponseReader<S>, (RequestWriter<S>, IoError)>

Send the request and construct a ResponseReader out of it.

If the request sending fails in any way, a condition will be raised; if handled, the original request will be returned as an Err.