Trait http::client::HttpSend

pub trait HttpSend {
    fn send_stream(&mut self, method: Method, path: &str, header: Option<&HashMap<String, String>>) -> Result<BufWriter<&mut Write>, Error>;
    fn get_reply(&mut self) -> Result<HttpReply<&mut Read>, Error>;

    fn send(&mut self, method: Method, path: &str, header: Option<&HashMap<String, String>>, data: Option<&[u8]>) -> Result<HttpReply<&mut Read>, Error> { ... }
}

Trait for object capable of sending HttpRequests

Required Methods

fn send_stream(&mut self, method: Method, path: &str, header: Option<&HashMap<String, String>>) -> Result<BufWriter<&mut Write>, Error>

Start a new request and return BufWriter to the underlying stream so you can write the request body.

When done, don't forget to call flush() on the BufWriter in order to flush all the buffer

fn get_reply(&mut self) -> Result<HttpReply<&mut Read>, Error>

Get the reply from stream. Must be called only after a request has been sent

Provided Methods

fn send(&mut self, method: Method, path: &str, header: Option<&HashMap<String, String>>, data: Option<&[u8]>) -> Result<HttpReply<&mut Read>, Error>

Send a full request and return the HttpReply.

If some data are provided, they are written to the request body, and the corresponding Content-Lenth header is inserted into request's properties