Function http::common::read_hexadecimal

pub fn read_hexadecimal<R: Reader, N: Unsigned + NumCast + PartialOrd + CheckedMul + CheckedAdd>(reader: &mut R, expected_end: |u8| -> bool) -> IoResult<N>

Read a positive hexadecimal integer from the given reader.

Arguments

reader - the reader to read the hexadecimal digits from (and whatever byte comes next)

expected_end - this function is called with the byte that follows the last hexadecimal digit and should return true if that byte is what is expected to follow the number.

Return value

None, if the number overflows;

None, if there is a leading zero;

None, if all the digits are read and the expected_end function is false

Should everything work as designed (i.e. none of these conditions occur) a Some is returned.