Trait http::connecter::Connecter

pub trait Connecter {
    fn connect(addr: SocketAddr, host: &str, use_ssl: bool) -> IoResult<Self>;
}

A trait for the concept of opening a stream connected to a IP socket address.

Why is this here? So that we can implement things which must make connections in terms of anything that can make such a connection rather than in terms of TcpStream only. This is handy for testing and for SSL.

Required Methods

fn connect(addr: SocketAddr, host: &str, use_ssl: bool) -> IoResult<Self>