Enum http::header::Entry

pub enum Entry<'a, T: 'a> {
    Occupied(OccupiedEntry<'a, T>),
    Vacant(VacantEntry<'a, T>),
}

A view into a single location in a HeaderMap, which may be vacant or occupied.

Variants
Occupied(OccupiedEntry<'a, T>)

An occupied entry

Vacant(VacantEntry<'a, T>)

A vacant entry

Methods

impl<'a, T> Entry<'a, T>

fn or_insert(self, default: T) -> &'a mut T

Ensures a value is in the entry by inserting the default if empty.

Returns a mutable reference to the first value in the entry.

Examples

let mut map = HeaderMap::new();

let headers = &[
    "content-length",
    "x-hello",
    "Content-Length",
    "x-world",
];

for &header in headers {
    let counter = map.entry(header).or_insert(0);
    *counter += 1;
}

assert_eq!(map["content-length"], 2);
assert_eq!(map["x-hello"], 1);

fn or_insert_with<F: FnOnce() -> T>(self, default: F) -> &'a mut T

Ensures a value is in the entry by inserting the result of the default function if empty.

The default function is not called if the entry exists in the map. Returns a mutable reference to the first value in the entry.

Examples

Basic usage.

let mut map = HeaderMap::new();

let res = map.entry("x-hello")
    .or_insert_with(|| "world".to_string());

assert_eq!(res, "world");

The default function is not called if the entry exists in the map.

let mut map = HeaderMap::new();
map.insert("x-hello", "world".to_string());

let res = map.entry("x-hello")
    .or_insert_with(|| unreachable!());

assert_eq!(res, "world");

fn key(&self) -> &HeaderName

Returns a reference to the entry's key

Examples

let mut map: HeaderMap<i32> = HeaderMap::new();

assert_eq!(map.entry("x-hello").key().as_str(), "x-hello");