Struct http::header::GetAll

pub struct GetAll<'a, T: 'a> { /* fields omitted */ }

A view to all values stored in a single entry.

This struct is returned by HeaderMap::get_all.

Methods

impl<'a, T> GetAll<'a, T>

fn key(&self) -> &HeaderName

Returns a reference to the entry's key.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

assert_eq!("x-hello", map.get_all("x-hello").unwrap().key().as_str());

fn get(&self) -> &T

Get a reference to the first value in the set.

Values are stored in insertion order.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

assert_eq!(
    map.get_all("x-hello").unwrap().get(),
    &"world");

map.append("x-hello", "earth");

assert_eq!(
    map.get_all("x-hello").unwrap().get(),
    &"world");

fn iter(&self) -> ValueIter<T>

Returns an iterator visiting all values associated with the entry.

Values are iterated in insertion order.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");
map.append("x-hello", "earth");

let values = map.get_all("x-hello").unwrap();
let mut iter = values.iter();
assert_eq!(&"world", iter.next().unwrap());
assert_eq!(&"earth", iter.next().unwrap());
assert!(iter.next().is_none());