Struct http::header::OccupiedEntry

pub struct OccupiedEntry<'a, T: 'a> { /* fields omitted */ }

A view into a single occupied location in a HeaderMap.

This struct is returned as part of the Entry enum.

Methods

impl<'a, T> OccupiedEntry<'a, T>

fn key(&self) -> &HeaderName

Returns a reference to the entry's key.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(e) = map.entry("x-hello") {
    assert_eq!("x-hello", e.key().as_str());
}

fn get(&self) -> &T

Get a reference to the first value in the entry.

Values are stored in insertion order.

Panics

get panics if there are no values associated with the entry.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    assert_eq!(e.get(), &"world");

    e.append("earth");

    assert_eq!(e.get(), &"world");
}

fn get_mut(&mut self) -> &mut T

Get a mutable reference to the first value in the entry.

Values are stored in insertion order.

Panics

get_mut panics if there are no values associated with the entry.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world".to_string());

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    e.get_mut().push_str("-2");
    assert_eq!(e.get(), &"world-2");
}

fn into_mut(self) -> &'a mut T

Converts the OccupiedEntry into a mutable reference to the first value.

The lifetime of the returned reference is bound to the original map.

Panics

into_mut panics if there are no values associated with the entry.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world".to_string());
map.append("x-hello", "earth".to_string());

if let Entry::Occupied(e) = map.entry("x-hello") {
    e.into_mut().push_str("-2");
}

assert_eq!("world-2", map["x-hello"]);

fn insert(&mut self, value: T) -> T

Sets the value of the entry.

All previous values associated with the entry are removed and the first one is returned. See insert_mult for an API that returns all values.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    let mut prev = e.insert("earth");
    assert_eq!("world", prev);
}

assert_eq!("earth", map["x-hello"]);

fn insert_mult(&mut self, value: T) -> ValueDrain<T>

Sets the value of the entry.

This function does the same as insert except it returns an iterator that yields all values previously associated with the key.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");
map.append("x-hello", "world2");

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    let mut prev = e.insert_mult("earth");
    assert_eq!("world", prev.next().unwrap());
    assert_eq!("world2", prev.next().unwrap());
    assert!(prev.next().is_none());
}

assert_eq!("earth", map["x-hello"]);

fn append(&mut self, value: T)

Insert the value into the entry.

The new value is appended to the end of the entry's value list. All previous values associated with the entry are retained.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    e.append("earth");
}

let values = map.get_all("x-hello").unwrap();
let mut i = values.iter();
assert_eq!("world", *i.next().unwrap());
assert_eq!("earth", *i.next().unwrap());

fn remove(self) -> T

Remove the entry from the map.

All values associated with the entry are removed and the first one is returned. See remove_entry_mult for an API that returns all values.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(e) = map.entry("x-hello") {
    let mut prev = e.remove();
    assert_eq!("world", prev);
}

assert!(!map.contains_key("x-hello"));

fn remove_entry(self) -> (HeaderName, T)

Remove the entry from the map.

The key and all values associated with the entry are removed and the first one is returned. See remove_entry_mult for an API that returns all values.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

if let Entry::Occupied(e) = map.entry("x-hello") {
    let (key, mut prev) = e.remove_entry();
    assert_eq!("x-hello", key.as_str());
    assert_eq!("world", prev);
}

assert!(!map.contains_key("x-hello"));

fn remove_entry_mult(self) -> (HeaderName, ValueDrain<'a, T>)

Remove the entry from the map.

The key and all values associated with the entry are removed and returned.

fn iter(&self) -> ValueIter<T>

Returns an iterator visiting all values associated with the entry.

Values are iterated in insertion order.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");
map.append("x-hello", "earth");

if let Entry::Occupied(e) = map.entry("x-hello") {
    let mut iter = e.iter();
    assert_eq!(&"world", iter.next().unwrap());
    assert_eq!(&"earth", iter.next().unwrap());
    assert!(iter.next().is_none());
}

fn iter_mut(&mut self) -> ValueIterMut<T>

Returns an iterator mutably visiting all values associated with the entry.

Values are iterated in insertion order.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world".to_string());
map.append("x-hello", "earth".to_string());

if let Entry::Occupied(mut e) = map.entry("x-hello") {
    for e in e.iter_mut() {
        e.push_str("-boop");
    }
}

let mut values = map.get_all("x-hello").unwrap();
let mut i = values.iter();
assert_eq!(&"world-boop", i.next().unwrap());
assert_eq!(&"earth-boop", i.next().unwrap());