Struct http::header::VacantEntry

pub struct VacantEntry<'a, T: 'a> { /* fields omitted */ }

A view into a single empty location in a HeaderMap.

This struct is returned as part of the Entry enum.

Methods

impl<'a, T> VacantEntry<'a, T>

fn key(&self) -> &HeaderName

Returns a reference to the entry's key

Examples

let mut map: HeaderMap<i32> = HeaderMap::new();

assert_eq!(map.entry("x-hello").key().as_str(), "x-hello");

fn into_key(self) -> HeaderName

Take ownership of the key

Examples

let mut map: HeaderMap<i32> = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello") {
    assert_eq!(v.into_key().as_str(), "x-hello");
}

fn insert(self, value: T) -> &'a mut T

Insert the value into the entry.

The value will be associated with this entry's key. A mutable reference to the inserted value will be returned.

Examples

let mut map = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello") {
    v.insert("world");
}

assert_eq!(map["x-hello"], "world");

fn insert_entry(self, value: T) -> OccupiedEntry<'a, T>

Insert the value into the entry.

The value will be associated with this entry's key. The new OccupiedEntry is returned, allowing for further manipulation.

Examples

let mut map = HeaderMap::new();

if let Entry::Vacant(v) = map.entry("x-hello") {
    let mut e = v.insert_entry("world");
    e.insert("world2");
}

assert_eq!(map["x-hello"], "world2");