Trait http::header::IntoHeaderName

pub trait IntoHeaderName: Sealed { }

A marker trait used to identify values that can be used as insert keys to a HeaderMap.

Object Safety

This trait is not object safe.