Module http::headers

Types and utilities for working with headers in HTTP requests and responses.

This HTTP system is Special in that it uses lots of strong typing for its header system. All known HTTP headers are type checked, rather than being dealt with as strings all the time. Only unknown headers are stored in a map in the traditional way.

Modules

accept_ranges
The Accept-Ranges request header, defined in RFC 2616, Section 14.5.

connection
The Connection general header, defined in RFC 2616, Section 14.10.

content_type
The Content-Type entity header, defined in RFC 2616, Section 14.17.

etag

host
The Host request header, defined in RFC 2616, Section 14.23.

request

response

serialization_utils
Utility functions for assisting with conversion of headers from and to the HTTP text form.

test_utils

transfer_encoding
The Transfer-Encoding request header, defined in RFC 2616, sections 14.41 and 3.6.

Structs

HeaderValueByteIterator
An iterator over the bytes of a header value. This ensures one cannot read past the end of a header mistakenly and that linear white space is handled correctly so that nothing else needs to worry about it. Any linear whitespace (multiple spaces outside of a quoted-string) is compacted into a single SP.

Enums

ConsumeCommaLWSResult

HeaderLineErr

Traits

CommaListHeaderConvertible
A header with multiple comma-separated values. Implement this and a HeaderConvertible implementation for Vec is yours for free—just make sure your reading does not consume the comma.

HeaderConvertible
A datatype for headers.

HeaderEnum

Functions

header_enum_from_stream
Shifted out of being a default method to fix an ICE (not yet reported, TODO)

Type Definitions

DeltaSeconds