Struct http::headers::request::HeaderCollection

pub struct HeaderCollection {
    pub cache_control: Option<String>,
    pub connection: Option<Vec<Connection>>,
    pub date: Option<Tm>,
    pub pragma: Option<String>,
    pub trailer: Option<String>,
    pub transfer_encoding: Option<Vec<TransferCoding>>,
    pub upgrade: Option<String>,
    pub via: Option<String>,
    pub warning: Option<String>,
    pub accept: Option<String>,
    pub accept_charset: Option<String>,
    pub accept_encoding: Option<String>,
    pub accept_language: Option<String>,
    pub authorization: Option<String>,
    pub expect: Option<String>,
    pub from: Option<String>,
    pub host: Option<Host>,
    pub if_match: Option<String>,
    pub if_modified_since: Option<Tm>,
    pub if_none_match: Option<String>,
    pub if_range: Option<String>,
    pub if_unmodified_since: Option<Tm>,
    pub max_forwards: Option<uint>,
    pub proxy_authorization: Option<String>,
    pub range: Option<String>,
    pub referer: Option<String>,
    pub te: Option<String>,
    pub user_agent: Option<String>,
    pub allow: Option<Vec<Method>>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: Option<uint>,
    pub content_location: Option<String>,
    pub content_md5: Option<String>,
    pub content_range: Option<String>,
    pub content_type: Option<MediaType>,
    pub expires: Option<Tm>,
    pub last_modified: Option<Tm>,
    pub extensions: TreeMap<String, String>,
}

Fields

cache_control

connection

date

pragma

trailer

transfer_encoding

upgrade

via

warning

accept

accept_charset

accept_encoding

accept_language

authorization

expect

from

host

if_match

if_modified_since

if_none_match

if_range

if_unmodified_since

max_forwards

proxy_authorization

range

referer

te

user_agent

allow

content_encoding

content_language

content_length

content_location

content_md5

content_range

content_type

expires

last_modified

extensions

Methods

impl HeaderCollection

fn new() -> HeaderCollection

fn insert(&mut self, header: Header)

Consume a header, putting it into this structure.

fn insert_raw(&mut self, name: String, value: &[u8]) -> Result<(), ()>

Insert a raw header into the collection. This will return an error if the value is not valid UTF-8 or if the name is that of a strongly-typed header and the value is not a valid value for that header.

fn iter<'a>(&'a self) -> HeaderCollectionIterator<'a>

fn write_all<W: Writer>(&self, writer: &mut W) -> IoResult<()>

Write all the headers to a writer. This includes an extra \r\n at the end to signal end of headers.