Struct http::headers::response::HeaderCollection

pub struct HeaderCollection {
    pub cache_control: Option<String>,
    pub connection: Option<Vec<Connection>>,
    pub date: Option<Tm>,
    pub pragma: Option<String>,
    pub trailer: Option<String>,
    pub transfer_encoding: Option<Vec<TransferCoding>>,
    pub upgrade: Option<String>,
    pub via: Option<String>,
    pub warning: Option<String>,
    pub accept_patch: Option<String>,
    pub accept_ranges: Option<AcceptableRanges>,
    pub age: Option<String>,
    pub etag: Option<EntityTag>,
    pub location: Option<Url>,
    pub proxy_authenticate: Option<String>,
    pub retry_after: Option<String>,
    pub server: Option<String>,
    pub vary: Option<String>,
    pub www_authenticate: Option<String>,
    pub allow: Option<Vec<Method>>,
    pub content_encoding: Option<String>,
    pub content_language: Option<String>,
    pub content_length: Option<uint>,
    pub content_location: Option<String>,
    pub content_md5: Option<String>,
    pub content_range: Option<String>,
    pub content_type: Option<MediaType>,
    pub expires: Option<String>,
    pub last_modified: Option<Tm>,
    pub extensions: TreeMap<String, String>,
}

Fields

cache_control

connection

date

pragma

trailer

transfer_encoding

upgrade

via

warning

accept_patch

accept_ranges

age

etag

location

proxy_authenticate

retry_after

server

vary

www_authenticate

allow

content_encoding

content_language

content_length

content_location

content_md5

content_range

content_type

expires

last_modified

extensions

Methods

impl HeaderCollection

fn new() -> HeaderCollection

fn insert(&mut self, header: Header)

Consume a header, putting it into this structure.

fn insert_raw(&mut self, name: String, value: &[u8]) -> Result<(), ()>

Insert a raw header into the collection. This will return an error if the value is not valid UTF-8 or if the name is that of a strongly-typed header and the value is not a valid value for that header.

fn iter<'a>(&'a self) -> HeaderCollectionIterator<'a>

fn write_all<W: Writer>(&self, writer: &mut W) -> IoResult<()>

Write all the headers to a writer. This includes an extra \r\n at the end to signal end of headers.