Module http::headers::serialization_utils

Utility functions for assisting with conversion of headers from and to the HTTP text form.

Traits

WriterUtil

Functions

comma_join
Join a vector of values with commas, as is common for HTTP headers.

comma_split
Split a value on commas, as is common for HTTP headers.

comma_split_iter

maybe_quoted_string
Make a string into a ( token | quoted-string ), preferring a token

maybe_unquote_string
Parse a ( token | quoted-string ). Returns None if it is not valid.

normalise_header_name
Normalise an HTTP header name.

push_maybe_quoted_string
Push a ( token | quoted-string ) onto a string and return it again

push_parameter

push_parameters

push_quoted_string
Quote a string, to turn it into an RFC 2616 quoted-string

quoted_string
Quote a string, to turn it into an RFC 2616 quoted-string

unquote_string
Parse a quoted-string. Returns None if the string is not a valid quoted-string.