Struct http::headers::HeaderValueByteIterator

pub struct HeaderValueByteIterator<'a, R> {
    pub reader: &'a mut R,
    pub next_byte: Option<u8>,
    pub at_start: bool,
    // some fields omitted
}

An iterator over the bytes of a header value. This ensures one cannot read past the end of a header mistakenly and that linear white space is handled correctly so that nothing else needs to worry about it. Any linear whitespace (multiple spaces outside of a quoted-string) is compacted into a single SP.

Fields

reader

next_byte

This field serves two purposes. During iteration, it will typically be None, but certain cases will cause it to be a Some, meaning that the next next() call will return that value rather than reading a new byte. At the end of iteration (after next() has returned None), it will be the extra byte which it has had to consume from the stream because of the possibility of linear white space of the form CR LF SP. It is guaranteed that if self.state == Finished this will be a Some.

at_start

Methods

impl<'a, R: Reader> HeaderValueByteIterator<'a, R>

fn new(reader: &'a mut R) -> HeaderValueByteIterator<'a, R>

fn verify_consumed(&mut self) -> bool

Check that the entire header value has been consumed.

Should there be any trailing linear white space, it is dropped.

Be cautious using this function as it is destructive, losing a character in the case where the value has not been entirely consumed.

This should only be called when finished with a value and ensuring that there aren't unexpected characters

Suggested usage is in a from_stream method::

if reader.verify_consumed() {
    Some(header)
} else {
    None
}

... however, this common case is handled with the some_if_consumed method, so you may very well not need to call this function directly.

fn some_if_consumed<T>(&mut self, t: T) -> Option<T>

Turn a constructed header value into an Option: Some(value) if the header value is consumed or None if it is not, thus indicating: "I'm finished and expect nothing more. Anything more is an error."

fn consume_optional_lws(&mut self)

Consume optional *LWS. That is, zero or more of SP and HT, until it gets to something other than SP and HT or gets to the end of the header.

fn consume_lws(&mut self) -> bool

Consume 1*LWS. That is, one or more of SP and HT, until it gets to something other than SP and HT or gets to the end of the header.

fn consume_comma_lws(&mut self) -> ConsumeCommaLWSResult

Return values: - CommaConsumed if there was a comma and it was consumed; - EndOfValue if the header value has been completely consumed; - ErrCommaNotFound if the next thing wasn't a comma (this is an error state)

fn read_quoted_string(&mut self, already_opened: bool) -> Option<String>

Read a quoted-string from the current position. If the quoted-string is not begun immediately or the header ends before it is completed, then None is returned; TODO: decide if I can return the bytes read (at present, escapes and double quote would be lost if I did that).

fn read_token_or_quoted_string(&mut self) -> Option<String>

Read a token (RFC 2616 definition) from the header value.

If no token begins at the current point of the header, None will also be returned.

fn read_token(&mut self) -> Option<String>

Read a token (RFC 2616 definition) from the header value.

If no token begins at the current point of the header, None will also be returned.