Struct http::messages::HttpReply

pub struct HttpReply<T: Read> {
    // some fields omitted
}

A structure that represents an HTTP reply

It contains an already parsed header information, and offers a BufReader<T> to read reply content

Methods

impl<T: Read> HttpReply<T>

fn parse(reader: BufReader<T>) -> Result<HttpReply<T>, Error>

Contruct a new HttpReply by parsing the input from reader

Examples

use std::net::TcpStream;
use std::io::BufReader;
use http::messages::HttpReply;
let mut socket = TcpStream::connect("host_address:port").unwrap();
// Do some stuff with socket and assume an http reply is coming
let reader = BufReader::new(socket);
let r = HttpReply::parse(reader).unwrap();

fn get_length(&self) -> Result<usize, Error>

Get the Content-Length property from header

fn get_reader(&mut self) -> &mut BufReader<T>

Return a BufReader to read the reply's content

fn get_version(&self) -> &String

Get the HTTP version from reply. Returns a string like "HTTP/1.0"

fn get_code(&self) -> u32

Get the status code from reply

fn get_status(&self) -> &String

Get the status text from reply

fn get_property(&self, key: &String) -> Option<&String>

Get a property from reply header

fn get_properties_name(&self) -> Keys<String, String>

Return an iterator over properties names set in reply header

fn iter(&self) -> Iter<String, String>

Return an iterator over properties from reply header

fn read_all(&mut self) -> Result<Vec<u8>, Error>

Read all the reply content into a buffer and return this buffer

fn read_string(&mut self) -> Result<String, Error>

Read all the repl and return it into a String