Module http::method

The HTTP request method

This module contains HTTP-method related structs and errors and such. The main type of this module, Method, is also reexported at the root of the crate as http::Method and is intended for import through that location primarily.

This module also contains constants for a number of common HTTP methods such as GET, POST, etc.

Examples

use http::{method, Method};

assert_eq!(method::GET, Method::from_bytes(b"GET").unwrap());
assert!(method::GET.is_idempotent());
assert_eq!(method::POST.as_str(), "POST");

Structs

InvalidMethod
A possible error value when converting Method from bytes.

Method
The Request Method (VERB)

Constants

CONNECT
CONNECT

DELETE
DELETE

GET
GET

HEAD
HEAD

OPTIONS
OPTIONS

PATCH
PATCH

POST
POST

PUT
PUT

TRACE
TRACE