Struct http::method::Method

pub struct Method(_);

The Request Method (VERB)

This type also contains constants for a number of common HTTP methods such as GET, POST, etc.

Currently includes 8 variants representing the 8 methods defined in RFC 7230, plus PATCH, and an Extension variant for all extensions.

Examples

use http::Method;

assert_eq!(Method::GET, Method::from_bytes(b"GET").unwrap());
assert!(Method::GET.is_idempotent());
assert_eq!(Method::POST.as_str(), "POST");

Implementations

impl Method

pub const GET: Method = _

GET

pub const POST: Method = _

POST

pub const PUT: Method = _

PUT

pub const DELETE: Method = _

DELETE

pub const HEAD: Method = _

HEAD

pub const OPTIONS: Method = _

OPTIONS

pub const CONNECT: Method = _

CONNECT

pub const PATCH: Method = _

PATCH

pub const TRACE: Method = _

TRACE

pub fn from_bytes(src: &[u8]) -> Result<Method, InvalidMethod>

Converts a slice of bytes to an HTTP method.

pub fn is_safe(&self) -> bool

Whether a method is considered “safe”, meaning the request is essentially read-only.

See the spec for more words.

pub fn is_idempotent(&self) -> bool

Whether a method is considered “idempotent”, meaning the request has the same result if executed multiple times.

See the spec for more words.

pub fn as_str(&self) -> &str

Return a &str representation of the HTTP method