Struct http::request::Builder

pub struct Builder { /* private fields */ }

An HTTP request builder

This type can be used to construct an instance or Request through a builder-like pattern.

Implementations

impl Builder

pub fn new() -> Builder

Creates a new default instance of Builder to construct a Request.

Examples

let req = request::Builder::new()
    .method("POST")
    .body(())
    .unwrap();

pub fn method<T>(self, method: T) -> Builder
where Method: TryFrom<T>, <Method as TryFrom<T>>::Error: Into<Error>,

Set the HTTP method for this request.

By default this is GET.

Examples

let req = Request::builder()
    .method("POST")
    .body(())
    .unwrap();

pub fn method_ref(&self) -> Option<&Method>

Get the HTTP Method for this request.

By default this is GET. If builder has error, returns None.

Examples

let mut req = Request::builder();
assert_eq!(req.method_ref(),Some(&Method::GET));

req = req.method("POST");
assert_eq!(req.method_ref(),Some(&Method::POST));

pub fn uri<T>(self, uri: T) -> Builder
where Uri: TryFrom<T>, <Uri as TryFrom<T>>::Error: Into<Error>,

Set the URI for this request.

By default this is /.

Examples

let req = Request::builder()
    .uri("https://www.rust-lang.org/")
    .body(())
    .unwrap();

pub fn uri_ref(&self) -> Option<&Uri>

Get the URI for this request

By default this is /.

Examples

let mut req = Request::builder();
assert_eq!(req.uri_ref().unwrap(), "/" );

req = req.uri("https://www.rust-lang.org/");
assert_eq!(req.uri_ref().unwrap(), "https://www.rust-lang.org/" );

pub fn version(self, version: Version) -> Builder

Set the HTTP version for this request.

By default this is HTTP/1.1

Examples

let req = Request::builder()
    .version(Version::HTTP_2)
    .body(())
    .unwrap();

pub fn version_ref(&self) -> Option<&Version>

Get the HTTP version for this request

By default this is HTTP/1.1.

Examples

let mut req = Request::builder();
assert_eq!(req.version_ref().unwrap(), &Version::HTTP_11 );

req = req.version(Version::HTTP_2);
assert_eq!(req.version_ref().unwrap(), &Version::HTTP_2 );

pub fn header<K, V>(self, key: K, value: V) -> Builder
where HeaderName: TryFrom<K>, <HeaderName as TryFrom<K>>::Error: Into<Error>, HeaderValue: TryFrom<V>, <HeaderValue as TryFrom<V>>::Error: Into<Error>,

Appends a header to this request builder.

This function will append the provided key/value as a header to the internal HeaderMap being constructed. Essentially this is equivalent to calling HeaderMap::append.

Examples

let req = Request::builder()
    .header("Accept", "text/html")
    .header("X-Custom-Foo", "bar")
    .body(())
    .unwrap();

pub fn headers_ref(&self) -> Option<&HeaderMap<HeaderValue>>

Get header on this request builder. when builder has error returns None

Example

let req = Request::builder()
    .header("Accept", "text/html")
    .header("X-Custom-Foo", "bar");
let headers = req.headers_ref().unwrap();
assert_eq!( headers["Accept"], "text/html" );
assert_eq!( headers["X-Custom-Foo"], "bar" );

pub fn headers_mut(&mut self) -> Option<&mut HeaderMap<HeaderValue>>

Get headers on this request builder.

When builder has error returns None.

Example

let mut req = Request::builder();
{
  let headers = req.headers_mut().unwrap();
  headers.insert("Accept", HeaderValue::from_static("text/html"));
  headers.insert("X-Custom-Foo", HeaderValue::from_static("bar"));
}
let headers = req.headers_ref().unwrap();
assert_eq!( headers["Accept"], "text/html" );
assert_eq!( headers["X-Custom-Foo"], "bar" );

pub fn extension<T>(self, extension: T) -> Builder
where T: Any + Send + Sync + 'static,

Adds an extension to this builder

Examples

let req = Request::builder()
    .extension("My Extension")
    .body(())
    .unwrap();

assert_eq!(req.extensions().get::<&'static str>(),
           Some(&"My Extension"));

pub fn extensions_ref(&self) -> Option<&Extensions>

Get a reference to the extensions for this request builder.

If the builder has an error, this returns None.

Example

let req = Request::builder().extension("My Extension").extension(5u32);
let extensions = req.extensions_ref().unwrap();
assert_eq!(extensions.get::<&'static str>(), Some(&"My Extension"));
assert_eq!(extensions.get::<u32>(), Some(&5u32));

pub fn extensions_mut(&mut self) -> Option<&mut Extensions>

Get a mutable reference to the extensions for this request builder.

If the builder has an error, this returns None.

Example

let mut req = Request::builder().extension("My Extension");
let mut extensions = req.extensions_mut().unwrap();
assert_eq!(extensions.get::<&'static str>(), Some(&"My Extension"));
extensions.insert(5u32);
assert_eq!(extensions.get::<u32>(), Some(&5u32));

pub fn body<T>(self, body: T) -> Result<Request<T>>

“Consumes” this builder, using the provided body to return a constructed Request.

Errors

This function may return an error if any previously configured argument failed to parse or get converted to the internal representation. For example if an invalid head was specified via header("Foo", "Bar\r\n") the error will be returned when this function is called rather than when header was called.

Examples

let request = Request::builder()
    .body(())
    .unwrap();