Struct http::response::Response

pub struct Response<T> { /* fields omitted */ }

Represents an HTTP response

An HTTP response consists of a head and a potentially optional body. The body component is generic, enabling arbitrary types to represent the HTTP body. For example, the body could be Vec<u8>, a Stream of byte chunks, or a value that has been deserialized.

Typically you'll work with responses on the client side as the result of sending a Request and on the server you'll be generating a Request to send back to the client.

Examples

Creating a Response to return

use http::{Request, Response};
use http::status::OK;

fn respond_to(req: Request<()>) -> http::Result<Response<()>> {
    let mut response = Response::builder();
    response.header("Foo", "Bar")
            .status(OK);

    if req.headers().contains_key("Another-Header") {
        response.header("Another-Header", "Ack");
    }

    response.body(())
}

A simple 404 handler

use http::{Request, Response};
use http::status::NOT_FOUND;

fn not_found(_req: Request<()>) -> http::Result<Response<()>> {
    Response::builder()
        .status(NOT_FOUND)
        .body(())
}

Or otherwise inspecting the result of a request:

use http::{Request, Response};
use http::status::NOT_FOUND;

fn get(url: &str) -> http::Result<Response<()>> {
    // ...
}

let response = get("https://www.rust-lang.org/").unwrap();

if !response.status().is_success() {
    panic!("failed to get a successful response status!");
}

if let Some(date) = response.headers().get("Date") {
    // we've got a `Date` header!
}

let body = response.body();
// ...

Deserialize a response of bytes via json:

use http::Response;
use serde::de;

fn deserialize<T>(req: Response<Vec<u8>>) -> serde_json::Result<Response<T>>
    where for<'de> T: de::Deserialize<'de>,
{
    let (parts, body) = req.into_parts();
    let body = serde_json::from_slice(&body)?;
    Ok(Response::from_parts(parts, body))
}

Or alternatively, serialize the body of a response to json

use http::Response;
use serde::ser;

fn serialize<T>(req: Response<T>) -> serde_json::Result<Response<Vec<u8>>>
    where T: ser::Serialize,
{
    let (parts, body) = req.into_parts();
    let body = serde_json::to_vec(&body)?;
    Ok(Response::from_parts(parts, body))
}

Methods

impl Response<()>

fn builder() -> Builder

Creates a new builder-style object to manufacture a Response

This method returns an instance of Builder which can be used to create a Response.

Examples

let response = Response::builder()
    .status(200)
    .header("X-Custom-Foo", "Bar")
    .body(())
    .unwrap();

impl<T> Response<T>

fn new(body: T) -> Response<T>

Creates a new blank Response with the body

The component ports of this response will be set to their default, e.g. the ok status, no headers, etc.

Examples

let response = Response::new("hello world");

assert_eq!(response.status(), status::OK);
assert_eq!(*response.body(), "hello world");

fn from_parts(parts: Parts, body: T) -> Response<T>

Creates a new Response with the given head and body

Examples

let response = Response::new("hello world");
let (mut parts, body) = response.into_parts();

parts.status = status::BAD_REQUEST;
let response = Response::from_parts(parts, body);

assert_eq!(response.status(), status::BAD_REQUEST);
assert_eq!(*response.body(), "hello world");

fn status(&self) -> StatusCode

Returns the StatusCode.

Examples

let response: Response<()> = Response::default();
assert_eq!(response.status(), status::OK);

fn status_mut(&mut self) -> &mut StatusCode

Returns a mutable reference to the associated StatusCode.

Examples

let mut response: Response<()> = Response::default();
*response.status_mut() = status::CREATED;
assert_eq!(response.status(), status::CREATED);

fn version(&self) -> Version

Returns a reference to the associated version.

Examples

let response: Response<()> = Response::default();
assert_eq!(response.version(), version::HTTP_11);

fn version_mut(&mut self) -> &mut Version

Returns a mutable reference to the associated version.

Examples

let mut response: Response<()> = Response::default();
*response.version_mut() = version::HTTP_2;
assert_eq!(response.version(), version::HTTP_2);

fn headers(&self) -> &HeaderMap<HeaderValue>

Returns a reference to the associated header field map.

Examples

let response: Response<()> = Response::default();
assert!(response.headers().is_empty());

fn headers_mut(&mut self) -> &mut HeaderMap<HeaderValue>

Returns a mutable reference to the associated header field map.

Examples

let mut response: Response<()> = Response::default();
response.headers_mut().insert("hello", HeaderValue::from_static("world"));
assert!(!response.headers().is_empty());

fn extensions(&self) -> &Extensions

Returns a reference to the associated extensions.

Examples

let response: Response<()> = Response::default();
assert!(response.extensions().get::<i32>().is_none());

fn extensions_mut(&mut self) -> &mut Extensions

Returns a mutable reference to the associated extensions.

Examples

let mut response: Response<()> = Response::default();
response.extensions_mut().insert("hello");
assert_eq!(response.extensions().get(), Some(&"hello"));

fn body(&self) -> &T

Returns a reference to the associated HTTP body.

Examples

let response: Response<String> = Response::default();
assert!(response.body().is_empty());

fn body_mut(&mut self) -> &mut T

Returns a mutable reference to the associated HTTP body.

Examples

let mut response: Response<String> = Response::default();
response.body_mut().push_str("hello world");
assert!(!response.body().is_empty());

fn into_parts(self) -> (Parts, T)

Consumes the response returning the head and body parts.

Examples

let response: Response<()> = Response::default();
let (parts, body) = response.into_parts();
assert_eq!(parts.status, status::OK);