Struct http::server::request::Request

pub struct Request {
    pub remote_addr: Option<SocketAddr>,
    pub headers: Box<HeaderCollection>,
    pub body: String,
    pub method: Method,
    pub request_uri: RequestUri,
    pub close_connection: bool,
    pub version: (uint, uint),
}

An HTTP request sent to the server.

Fields

remote_addr

The originating IP address of the request.

headers

The host name and IP address that the request was sent to; this must always be specified for HTTP/1.1 requests (or the request will be rejected), but for HTTP/1.0 requests the Host header was not defined, and so this field will probably be None in such cases. The headers sent with the request.

body

The body of the request; empty for such methods as GET.

method

The HTTP method for the request.

request_uri

The URI that was requested, as found in the Request-URI of the Request-Line. You will almost never need to use this; you should prefer the url field instead.

close_connection

Whether to close the TCP connection when the request has been served. The alternative is keeping the connection open and waiting for another request.

version

The HTTP version number; typically (1, 1) or, less commonly, (1, 0).

Methods

impl Request

fn load(stream: &mut BufferedStream<TcpStream>) -> (Request, Result<(), Status>)

Get a response from an open socket.