Enum http::server::request::RequestUri

pub enum RequestUri {
    Star,
    AbsoluteUri(Url),
    AbsolutePath(String),
    Authority(String),
}

The URI (Request-URI in RFC 2616) as specified in the Status-Line of an HTTP request

Variants

Star

'The asterisk "*" means that the request does not apply to a particular resource, but to the server itself, and is only allowed when the method used does not necessarily apply to a resource. One example would be "OPTIONS * HTTP/1.1" '

AbsoluteUri

'The absoluteURI form is REQUIRED when the request is being made to a proxy. The proxy is requested to forward the request or service it from a valid cache, and return the response. Note that the proxy MAY forward the request on to another proxy or directly to the server specified by the absoluteURI. In order to avoid request loops, a proxy MUST be able to recognize all of its server names, including any aliases, local variations, and the numeric IP address. An example Request-Line would be: "GET http://www.w3.org/pub/WWW/TheProject.html HTTP/1.1"'

AbsolutePath

'To allow for transition to absoluteURIs in all requests in future versions of HTTP, all HTTP/1.1 servers MUST accept the absoluteURI form in requests, even though HTTP/1.1 clients will only generate them in requests to proxies.'

TODO: this shouldn't be a string; it should be further parsed. extra::net::url has some stuff which might help, but isn't public.

Authority

'The authority form is only used by the CONNECT method (CONNECT).'

TODO: this shouldn't be a string; it should be further parsed. extra::net::url has some stuff which might help, but isn't public.