Struct http::server::Config

pub struct Config {
    pub bind_address: SocketAddr,
}

The necessary configuration for an HTTP server.

At present, only the IP address and port to bind to are needed, but it's possible that other options may turn up later.

Fields

bind_address