Trait http::server::Server

pub trait Server: Send + Clone {
    fn handle_request(&self, request: Request, response: &mut ResponseWriter);
    fn get_config(&self) -> Config;

    fn serve_forever(self) { ... }
}

Required Methods

fn handle_request(&self, request: Request, response: &mut ResponseWriter)

fn get_config(&self) -> Config

Provided Methods

fn serve_forever(self)

Attempt to bind to the address and port and start serving forever.

This will only return if the initial connection fails or something else blows up.