Module http::status

HTTP status codes

This module contains HTTP-status code related structs an errors. The main type in this module is StatusCode which is not intended to be used through this module but rather the http::StatusCode type. This module also primarily contains a number of predefined constants for common HTTP status codes.

Examples

use http::StatusCode;

assert_eq!(StatusCode::from_u16(200).unwrap(), StatusCode::OK);
assert_eq!(StatusCode::NOT_FOUND.as_u16(), 404);
assert!(StatusCode::OK.is_success());

Structs

InvalidStatusCode
A possible error value when converting a StatusCode from a u16 or &str

StatusCode
An HTTP status code (status-code in RFC 7230 et al.).