Struct http::status::StatusCode

pub struct StatusCode(_);

An HTTP status code (status-code in RFC 7230 et al.).

This type contains constructor functions for all common status codes. It allows status codes in the range [0, 65535], as any u16 integer may be used as a status code for XHR requests. It is recommended to only use values between [100, 599], since only these are defined as valid status codes with a status class by HTTP.

IANA maintain the Hypertext Transfer Protocol (HTTP) Status Code Registry which is the source for this enum (with one exception, 418 I'm a teapot, which is inexplicably not in the register).

Examples

use http::{status, StatusCode};

assert_eq!(StatusCode::from_u16(200).unwrap(), status::OK);
assert_eq!(status::NOT_FOUND.as_u16(), 404);
assert!(status::OK.is_success());

Methods

impl StatusCode

fn from_u16(src: u16) -> Result<StatusCode, InvalidStatusCode>

Converts a u16 to a status code.

The function validates the correctness of the supplied u16. It must be greater or equal to 100 but less than 600.

Example

use http::{status, StatusCode};

let ok = StatusCode::from_u16(200).unwrap();
assert_eq!(ok, status::OK);

let err = StatusCode::from_u16(99);
assert!(err.is_err());

fn from_bytes(src: &[u8]) -> Result<StatusCode, InvalidStatusCode>

Converts a &[u8] to a status code

fn as_u16(&self) -> u16

Returns the u16 corresponding to this StatusCode.

Example

let status = http::status::OK;
assert_eq!(status.as_u16(), 200);

fn is_informational(&self) -> bool

Check if status is within 100-199.

fn is_success(&self) -> bool

Check if status is within 200-299.

fn is_redirection(&self) -> bool

Check if status is within 300-399.

fn is_client_error(&self) -> bool

Check if status is within 400-499.

fn is_server_error(&self) -> bool

Check if status is within 500-599.

impl StatusCode

fn canonical_reason(&self) -> Option<&'static str>

Get the standardised reason-phrase for this status code.

This is mostly here for servers writing responses, but could potentially have application at other times.

The reason phrase is defined as being exclusively for human readers. You should avoid deriving any meaning from it at all costs.

Bear in mind also that in HTTP/2.0 the reason phrase is abolished from transmission, and so this canonical reason phrase really is the only reason phrase you’ll find.

impl StatusCode

fn as_str(&self) -> &str

Returns a &str representation of the StatusCode

The return value only includes a numerical representation of the status code. The canonical reason is not included.