Struct http::Error

pub struct Error { /* fields omitted */ }

A generic "error" for HTTP connections

This error type is less specific than the error returned from other functions in this crate, but all other errors can be converted to this error. Consumers of this crate can typically consume and work with this form of error for conversions with the ? operator.