Struct http::uri::Authority

pub struct Authority { /* fields omitted */ }

Represents the authority component of a URI.

Methods

impl Authority

fn try_from_shared(s: Bytes) -> Result<Self, InvalidUriBytes>

Attempt to convert an Authority from Bytes.

This function will be replaced by a TryFrom implementation once the trait lands in stable.

Examples

extern crate bytes;

use bytes::Bytes;

let bytes = Bytes::from("example.com");
let authority = Authority::try_from_shared(bytes).unwrap();

assert_eq!(authority.host(), "example.com");

fn host(&self) -> &str

Get the host of this Authority.

The host subcomponent of authority is identified by an IP literal encapsulated within square brackets, an IPv4 address in dotted- decimal form, or a registered name. The host subcomponent is case-insensitive.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                        |---------|
                             |
                            host

Examples

let authority: Authority = "example.org:80".parse().unwrap();

assert_eq!(authority.host(), "example.org");

fn as_str(&self) -> &str

Return a str representation of the authority

fn into_bytes(self) -> Bytes

Converts this Authority back to a sequence of bytes