Struct http::uri::OriginForm

pub struct OriginForm { /* fields omitted */ }

Represents the path component of a URI

Methods

impl OriginForm

fn try_from_shared(src: Bytes) -> Result<Self, InvalidUriBytes>

Attempt to convert a OriginForm from Bytes.

This function will be replaced by a TryFrom implementation once the trait lands in stable.

Examples

extern crate bytes;

use bytes::Bytes;

let bytes = Bytes::from("/hello?world");
let origin_form = OriginForm::try_from_shared(bytes).unwrap();

assert_eq!(origin_form.path(), "/hello");
assert_eq!(origin_form.query(), Some("world"));

fn path(&self) -> &str

Returns the path component

The path component is case sensitive.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                                       |--------|
                                            |
                                          path

If the URI is * then the path component is equal to *.

Examples

let origin_form: OriginForm = "/hello/world".parse().unwrap();

assert_eq!(origin_form.path(), "/hello/world");

fn query(&self) -> Option<&str>

Returns the query string component

The query component contains non-hierarchical data that, along with data in the path component, serves to identify a resource within the scope of the URI's scheme and naming authority (if any). The query component is indicated by the first question mark ("?") character and terminated by a number sign ("#") character or by the end of the URI.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                                                  |-------------------|
                                                            |
                                                          query

Examples

With a query string component

let origin_form: OriginForm = "/hello/world?key=value&foo=bar".parse().unwrap();

assert_eq!(origin_form.query(), Some("key=value&foo=bar"));

Without a query string component

let origin_form: OriginForm = "/hello/world".parse().unwrap();

assert!(origin_form.query().is_none());

fn into_bytes(self) -> Bytes

Converts this OriginForm back to a sequence of bytes