Struct http::uri::PathAndQuery

pub struct PathAndQuery { /* fields omitted */ }

Represents the path component of a URI

Methods

impl PathAndQuery

pub fn from_shared(src: Bytes) -> Result<Self, InvalidUriBytes>

Attempt to convert a PathAndQuery from Bytes.

This function will be replaced by a TryFrom implementation once the trait lands in stable.

Examples

extern crate bytes;

use bytes::Bytes;

let bytes = Bytes::from("/hello?world");
let path_and_query = PathAndQuery::from_shared(bytes).unwrap();

assert_eq!(path_and_query.path(), "/hello");
assert_eq!(path_and_query.query(), Some("world"));

pub fn from_static(src: &'static str) -> Self

Convert a PathAndQuery from a static string.

This function will not perform any copying, however the string is checked to ensure that it is valid.

Panics

This function panics if the argument is an invalid path and query.

Examples

let v = PathAndQuery::from_static("/hello?world");

assert_eq!(v.path(), "/hello");
assert_eq!(v.query(), Some("world"));

pub fn path(&self) -> &str

Returns the path component

The path component is case sensitive.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                                       |--------|
                                            |
                                          path

If the URI is * then the path component is equal to *.

Examples

let path_and_query: PathAndQuery = "/hello/world".parse().unwrap();

assert_eq!(path_and_query.path(), "/hello/world");

pub fn query(&self) -> Option<&str>

Returns the query string component

The query component contains non-hierarchical data that, along with data in the path component, serves to identify a resource within the scope of the URI's scheme and naming authority (if any). The query component is indicated by the first question mark ("?") character and terminated by a number sign ("#") character or by the end of the URI.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                                                  |-------------------|
                                                            |
                                                          query

Examples

With a query string component

let path_and_query: PathAndQuery = "/hello/world?key=value&foo=bar".parse().unwrap();

assert_eq!(path_and_query.query(), Some("key=value&foo=bar"));

Without a query string component

let path_and_query: PathAndQuery = "/hello/world".parse().unwrap();

assert!(path_and_query.query().is_none());

pub fn as_str(&self) -> &str

Returns the path and query as a string component.

Examples

With a query string component

let path_and_query: PathAndQuery = "/hello/world?key=value&foo=bar".parse().unwrap();

assert_eq!(path_and_query.as_str(), "/hello/world?key=value&foo=bar");

Without a query string component

let path_and_query: PathAndQuery = "/hello/world".parse().unwrap();

assert_eq!(path_and_query.as_str(), "/hello/world");

pub fn into_bytes(self) -> Bytes

Converts this PathAndQuery back to a sequence of bytes