Crate http

Modules

buffer

client
Modules for making HTTP requests.

common
Poorly categorised functions for reading things used in multiple places.

connecter

headers
Types and utilities for working with headers in HTTP requests and responses.

method

rfc2616
Values taken from RFC 2616

server

status