Struct http::header::HeaderMap

pub struct HeaderMap<T> { /* fields omitted */ }

A set of HTTP headers

HeaderMap is an multimap of HeaderName to values.

Examples

Basic usage

let mut headers = HeaderMap::new();

headers.insert("Host", "example.com");
headers.insert("Content-Length", "123");

assert!(headers.contains_key("host"));
assert!(!headers.contains_key("Location"));

assert_eq!(headers["host"], "example.com");

headers.remove("host");

assert!(!headers.contains_key("host"));

Methods

impl<T> HeaderMap<T>

fn new() -> HeaderMap<T>

Create an empty HeaderMap.

The map will be created without any capacity. This function will not allocate.

Examples

let map: HeaderMap<u32> = HeaderMap::new();

assert!(map.is_empty());
assert_eq!(0, map.capacity());

fn with_capacity(capacity: usize) -> HeaderMap<T>

Create an empty HeaderMap with the specified capacity.

The returned map will allocate internal storage in order to hold about capacity elements without reallocating. However, this is a "best effort" as there are usage patterns that could cause additional allocations before capacity headers are stored in the map.

More capacity than requested may be allocated.

Examples

let map: HeaderMap<u32> = HeaderMap::with_capacity(10);

assert!(map.is_empty());
assert_eq!(12, map.capacity());

fn len(&self) -> usize

Returns the number of headers stored in the map.

This number represents the total number of values stored in the map. This number can be greater than or equal to the number of keys stored given that a single key may have more than one associated value.

Examples

let mut map = HeaderMap::new();

assert_eq!(0, map.len());

map.insert("x-header-one", "1");
map.insert("x-header-two", "2");

assert_eq!(2, map.len());

map.append("x-header-two", "deux");

assert_eq!(3, map.len());

fn keys_len(&self) -> usize

Returns the number of keys stored in the map.

This number will be less than or equal to len() as each key may have more than one associated value.

Examples

let mut map = HeaderMap::new();

assert_eq!(0, map.keys_len());

map.insert("x-header-one", "1");
map.insert("x-header-two", "2");

assert_eq!(2, map.keys_len());

map.insert("x-header-two", "deux");

assert_eq!(2, map.keys_len());

fn is_empty(&self) -> bool

Returns true if the map contains no elements.

Examples

let mut map = HeaderMap::new();

assert!(map.is_empty());

map.insert("x-hello", "world");

assert!(!map.is_empty());

fn clear(&mut self)

Clears the map, removing all key-value pairs. Keeps the allocated memory for reuse.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

map.clear();
assert!(map.is_empty());
assert!(map.capacity() > 0);

fn capacity(&self) -> usize

Returns the number of headers the map can hold without reallocating.

This number is an approximation as certain usage patterns could cause additional allocations before the returned capacity is filled.

Examples

let mut map = HeaderMap::new();

assert_eq!(0, map.capacity());

map.insert("x-hello", "world");
assert_eq!(6, map.capacity());

fn reserve(&mut self, additional: usize)

Reserves capacity for at least additional more headers to be inserted into the HeaderMap.

The header map may reserve more space to avoid frequent reallocations. Like with with_capacity, this will be a "best effort" to avoid allocations until additional more headers are inserted. Certain usage patterns could cause additional allocations before the number is reached.

Panics

Panics if the new allocation size overflows usize.

Examples

let mut map = HeaderMap::new();
map.reserve(10);

fn get<K: ?Sized>(&self, key: &K) -> Option<&T> where
K: HeaderMapKey,

Returns a reference to the value associated with the key.

If there are multiple values associated with the key, then the first one is returned. Use get_all to get all values associated with a given key. Returns None if there are no values associated with the key.

Examples

let mut map = HeaderMap::new();
assert!(map.get("x-hello").is_none());

map.insert("x-hello", "hello");
assert_eq!(map.get("x-hello").unwrap(), &"hello");

map.append("x-hello", "world");
assert_eq!(map.get("x-hello").unwrap(), &"hello");

fn get_mut<K: ?Sized>(&mut self, key: &K) -> Option<&mut T> where
K: HeaderMapKey,

Returns a mutable reference to the value associated with the key.

If there are multiple values associated with the key, then the first one is returned. Use entry to get all values associated with a given key. Returns None if there are no values associated with the key.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "hello".to_string());
map.get_mut("x-hello").unwrap().push_str("-world");

assert_eq!(map.get("x-hello").unwrap(), &"hello-world");

fn get_all<K: ?Sized>(&self, key: &K) -> Option<GetAll<T>> where
K: HeaderMapKey,

Returns a view of all values associated with a key.

The returned view does not incur any allocations and allows iterating the values associated with the key. See GetAll for more details. Returns None if there are no values associated with the key.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello");
map.append("x-hello", "goodbye");

let view = map.get_all("x-hello").unwrap();
assert_eq!(view.get(), &"hello");

let mut iter = view.iter();
assert_eq!(&"hello", iter.next().unwrap());
assert_eq!(&"goodbye", iter.next().unwrap());
assert!(iter.next().is_none());

fn contains_key<K: ?Sized>(&self, key: &K) -> bool where
K: HeaderMapKey,

Returns true if the map contains a value for the specified key.

Examples

let mut map = HeaderMap::new();
assert!(!map.contains_key("x-hello"));

map.insert("x-hello", "world");
assert!(map.contains_key("x-hello"));

fn iter(&self) -> Iter<T>

An iterator visiting all key-value pairs.

The iteration order is arbitrary, but consistent across platforms for the same crate version. Each key will be yielded once per associated value. So, if a key has 3 associated values, it will be yielded 3 times.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello");
map.append("x-hello", "goodbye");
map.insert("Content-Length", "123");

for (key, value) in map.iter() {
    println!("{:?}: {}", key, value);
}

fn iter_mut(&mut self) -> IterMut<T>

An iterator visiting all key-value pairs, with mutable value references.

The iterator order is arbitrary, but consistent across platforms for the same crate version. Each key will be yielded once per associated value, so if a key has 3 associated values, it will be yielded 3 times.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello".to_string());
map.append("x-hello", "goodbye".to_string());
map.insert("Content-Length", "123".to_string());

for (key, value) in map.iter_mut() {
    value.push_str("-boop");
}

fn keys(&self) -> Keys<T>

An iterator visiting all keys.

The iteration order is arbitrary, but consistent across platforms for the same crate version. Each key will be yielded only once even if it has multiple associated values.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello");
map.append("x-hello", "goodbye");
map.insert("Content-Length", "123");

for key in map.keys() {
    println!("{:?}", key);
}

fn values(&self) -> Values<T>

An iterator visiting all values.

The iteration order is arbitrary, but consistent across platforms for the same crate version.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello");
map.append("x-hello", "goodbye");
map.insert("Content-Length", "123");

for value in map.values() {
    println!("{}", value);
}

fn values_mut(&mut self) -> ValuesMut<T>

An iterator visiting all values mutably.

The iteration order is arbitrary, but consistent across platforms for the same crate version.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello".to_string());
map.append("x-hello", "goodbye".to_string());
map.insert("Content-Length", "123".to_string());

for value in map.values_mut() {
    value.push_str("-boop");
}

fn drain(&mut self) -> Drain<T>

Clears the map, returning all entries as an iterator.

The internal memory is kept for reuse.

Examples

let mut map = HeaderMap::new();

map.insert("x-hello", "hello");
map.append("x-hello", "goodbye");
map.insert("Content-Length", "123");

let mut drain = map.drain();

let (key, mut vals) = drain.next().unwrap();

assert_eq!("x-hello", key.as_str());
assert_eq!("hello", vals.next().unwrap());
assert_eq!("goodbye", vals.next().unwrap());
assert!(vals.next().is_none());

let (key, mut vals) = drain.next().unwrap();

assert_eq!("content-length", key.as_str());
assert_eq!("123", vals.next().unwrap());
assert!(vals.next().is_none());

fn entry<K>(&mut self, key: K) -> Entry<T> where
K: HeaderMapKey,

Gets the given key's corresponding entry in the map for in-place manipulation.

Examples

let mut map = HeaderMap::new();

let headers = &[
    "content-length",
    "x-hello",
    "Content-Length",
    "x-world",
];

for &header in headers {
    let counter = map.entry(header).or_insert(0);
    *counter += 1;
}

assert_eq!(map["content-length"], 2);
assert_eq!(map["x-hello"], 1);

fn insert<K>(&mut self, key: K, val: T) -> Option<T> where
K: HeaderMapKey,

Inserts a key-value pair into the map.

If the map did not previously have this key present, then None is returned.

If the map did have this key present, the new value is associated with the key and all previous values are removed. Note that only a single one of the previous values is returned. If there are multiple values that have been previously associated with the key, then the first one is returned. See insert_mult on OccupiedEntry for an API that returns all values.

The key is not updated, though; this matters for types that can be == without being identical.

Examples

let mut map = HeaderMap::new();
assert!(map.insert("x-hello", "world").is_none());
assert!(!map.is_empty());

let mut prev = map.insert("x-hello", "earth").unwrap();
assert_eq!("world", prev);

fn append<K>(&mut self, key: K, value: T) -> bool where
K: HeaderMapKey,

Inserts a key-value pair into the map.

If the map did not previously have this key present, then false is returned.

If the map did have this key present, the new value is pushed to the end of the list of values currently associated with the key. The key is not updated, though; this matters for types that can be == without being identical.

Examples

let mut map = HeaderMap::new();
assert!(map.insert("x-hello", "world").is_none());
assert!(!map.is_empty());

map.append("x-hello", "earth");

let values = map.get_all("x-hello").unwrap();
let mut i = values.iter();
assert_eq!("world", *i.next().unwrap());
assert_eq!("earth", *i.next().unwrap());

fn remove<K: ?Sized>(&mut self, key: &K) -> Option<T> where
K: HeaderMapKey,

Removes a key from the map, returning the value associated with the key.

Returns None if the map does not contain the key. If there are multiple values associated with the key, then the first one is returned. See remove_entry_mult on OccupiedEntry for an API that yields all values.

Examples

let mut map = HeaderMap::new();
map.insert("x-hello", "world");

let mut prev = map.remove("x-hello").unwrap();
assert_eq!("world", prev);

assert!(map.remove("x-hello").is_none());