Module http::status

HTTP status codes

This module contains HTTP-status code related structs an errors. The main type in this module is StatusCode which is not intended to be used through this module but rather the http::StatusCode type. This module also primarily contains a number of predefined constants for common HTTP status codes.

Examples

use http::{status, StatusCode};

assert_eq!(StatusCode::from_u16(200).unwrap(), status::OK);
assert_eq!(status::NOT_FOUND.as_u16(), 404);
assert!(status::OK.is_success());

Structs

InvalidStatusCode
A possible error value when converting a StatusCode from a u16 or &str

StatusCode
An HTTP status code (status-code in RFC 7230 et al.).

Constants

ACCEPTED
202 Accepted [RFC7231, Section 6.3.3]

ALREADY_REPORTED
208 Already Reported [RFC5842]

BAD_GATEWAY
502 Bad Gateway [RFC7231, Section 6.6.3]

BAD_REQUEST
400 Bad Request [RFC7231, Section 6.5.1]

CONFLICT
409 Conflict [RFC7231, Section 6.5.8]

CONTINUE
100 Continue [RFC7231, Section 6.2.1]

CREATED
201 Created [RFC7231, Section 6.3.2]

EXPECTATION_FAILED
417 Expectation Failed [RFC7231, Section 6.5.14]

FAILED_DEPENDENCY
424 Failed Dependency [RFC4918]

FORBIDDEN
403 Forbidden [RFC7231, Section 6.5.3]

FOUND
302 Found [RFC7231, Section 6.4.3]

GATEWAY_TIMEOUT
504 Gateway Timeout [RFC7231, Section 6.6.5]

GONE
410 Gone [RFC7231, Section 6.5.9]

HTTP_VERSION_NOT_SUPPORTED
505 HTTP Version Not Supported [RFC7231, Section 6.6.6]

IM_A_TEAPOT
418 I'm a teapot [curiously not registered by IANA but RFC2324]

IM_USED
226 IM Used [RFC3229]

INSUFFICIENT_STORAGE
507 Insufficient Storage [RFC4918]

INTERNAL_SERVER_ERROR
500 Internal Server Error [RFC7231, Section 6.6.1]

LENGTH_REQUIRED
411 Length Required [RFC7231, Section 6.5.10]

LOCKED
423 Locked [RFC4918]

LOOP_DETECTED
508 Loop Detected [RFC5842]

METHOD_NOT_ALLOWED
405 Method Not Allowed [RFC7231, Section 6.5.5]

MISDIRECTED_REQUEST
421 Misdirected Request RFC7540, Section 9.1.2

MOVED_PERMANENTLY
301 Moved Permanently [RFC7231, Section 6.4.2]

MULTIPLE_CHOICES
300 Multiple Choices [RFC7231, Section 6.4.1]

MULTI_STATUS
207 Multi-Status [RFC4918]

NETWORK_AUTHENTICATION_REQUIRED
511 Network Authentication Required [RFC6585]

NON_AUTHORITATIVE_INFORMATION
203 Non-Authoritative Information [RFC7231, Section 6.3.4]

NOT_ACCEPTABLE
406 Not Acceptable [RFC7231, Section 6.5.6]

NOT_EXTENDED
510 Not Extended [RFC2774]

NOT_FOUND
404 Not Found [RFC7231, Section 6.5.4]

NOT_IMPLEMENTED
501 Not Implemented [RFC7231, Section 6.6.2]

NOT_MODIFIED
304 Not Modified [RFC7232, Section 4.1]

NO_CONTENT
204 No Content [RFC7231, Section 6.3.5]

OK
200 OK [RFC7231, Section 6.3.1]

PARTIAL_CONTENT
206 Partial Content [RFC7233, Section 4.1]

PAYLOAD_TOO_LARGE
413 Payload Too Large [RFC7231, Section 6.5.11]

PAYMENT_REQUIRED
402 Payment Required [RFC7231, Section 6.5.2]

PERMANENT_REDIRECT
308 Permanent Redirect [RFC7238]

PRECONDITION_FAILED
412 Precondition Failed [RFC7232, Section 4.2]

PRECONDITION_REQUIRED
428 Precondition Required [RFC6585]

PROCESSING
102 Processing [RFC2518]

PROXY_AUTHENTICATION_REQUIRED
407 Proxy Authentication Required [RFC7235, Section 3.2]

RANGE_NOT_SATISFIABLE
416 Range Not Satisfiable [RFC7233, Section 4.4]

REQUEST_HEADER_FIELDS_TOO_LARGE
431 Request Header Fields Too Large [RFC6585]

REQUEST_TIMEOUT
408 Request Timeout [RFC7231, Section 6.5.7]

RESET_CONTENT
205 Reset Content [RFC7231, Section 6.3.6]

SEE_OTHER
303 See Other [RFC7231, Section 6.4.4]

SERVICE_UNAVAILABLE
503 Service Unavailable [RFC7231, Section 6.6.4]

SWITCHING_PROTOCOLS
101 Switching Protocols [RFC7231, Section 6.2.2]

TEMPORARY_REDIRECT
307 Temporary Redirect [RFC7231, Section 6.4.7]

TOO_MANY_REQUESTS
429 Too Many Requests [RFC6585]

UNAUTHORIZED
401 Unauthorized [RFC7235, Section 3.1]

UNAVAILABLE_FOR_LEGAL_REASONS
451 Unavailable For Legal Reasons [RFC7725]

UNPROCESSABLE_ENTITY
422 Unprocessable Entity [RFC4918]

UNSUPPORTED_MEDIA_TYPE
415 Unsupported Media Type [RFC7231, Section 6.5.13]

UPGRADE_REQUIRED
426 Upgrade Required [RFC7231, Section 6.5.15]

URI_TOO_LONG
414 URI Too Long [RFC7231, Section 6.5.12]

USE_PROXY
305 Use Proxy [RFC7231, Section 6.4.5]

VARIANT_ALSO_NEGOTIATES
506 Variant Also Negotiates [RFC2295]