Struct http::status::StatusCode

pub struct StatusCode(_);

An HTTP status code (status-code in RFC 7230 et al.).

This type contains constants for all common status codes. It allows status codes in the range [100, 599].

IANA maintain the Hypertext Transfer Protocol (HTTP) Status Code Registry which is the source for this enum (with one exception, 418 I'm a teapot, which is inexplicably not in the register).

Examples

use http::StatusCode;

assert_eq!(StatusCode::from_u16(200).unwrap(), StatusCode::OK);
assert_eq!(StatusCode::NOT_FOUND.as_u16(), 404);
assert!(StatusCode::OK.is_success());

Methods

impl StatusCode

pub fn from_u16(src: u16) -> Result<StatusCode, InvalidStatusCode>

Converts a u16 to a status code.

The function validates the correctness of the supplied u16. It must be greater or equal to 100 but less than 600.

Example

use http::StatusCode;

let ok = StatusCode::from_u16(200).unwrap();
assert_eq!(ok, StatusCode::OK);

let err = StatusCode::from_u16(99);
assert!(err.is_err());

pub fn from_bytes(src: &[u8]) -> Result<StatusCode, InvalidStatusCode>

Converts a &[u8] to a status code

pub fn as_u16(&self) -> u16

Returns the u16 corresponding to this StatusCode.

Note

This is the same as the From<StatusCode> implementation, but included as an inherent method because that implementation doesn't appear in rustdocs, as well as a way to force the type instead of relying on inference.

Example

let status = http::StatusCode::OK;
assert_eq!(status.as_u16(), 200);

pub fn as_str(&self) -> &str

Returns a &str representation of the StatusCode

The return value only includes a numerical representation of the status code. The canonical reason is not included.

Example

let status = http::StatusCode::OK;
assert_eq!(status.as_str(), "200");

pub fn canonical_reason(&self) -> Option<&'static str>

Get the standardised reason-phrase for this status code.

This is mostly here for servers writing responses, but could potentially have application at other times.

The reason phrase is defined as being exclusively for human readers. You should avoid deriving any meaning from it at all costs.

Bear in mind also that in HTTP/2.0 and HTTP/3.0 the reason phrase is abolished from transmission, and so this canonical reason phrase really is the only reason phrase you’ll find.

Example

let status = http::StatusCode::OK;
assert_eq!(status.canonical_reason(), Some("OK"));

pub fn is_informational(&self) -> bool

Check if status is within 100-199.

pub fn is_success(&self) -> bool

Check if status is within 200-299.

pub fn is_redirection(&self) -> bool

Check if status is within 300-399.

pub fn is_client_error(&self) -> bool

Check if status is within 400-499.

pub fn is_server_error(&self) -> bool

Check if status is within 500-599.

impl StatusCode

pub const CONTINUE: StatusCode

100 Continue [RFC7231, Section 6.2.1]

pub const SWITCHING_PROTOCOLS: StatusCode

101 Switching Protocols [RFC7231, Section 6.2.2]

pub const PROCESSING: StatusCode

102 Processing [RFC2518]

pub const OK: StatusCode

200 OK [RFC7231, Section 6.3.1]

pub const CREATED: StatusCode

201 Created [RFC7231, Section 6.3.2]

pub const ACCEPTED: StatusCode

202 Accepted [RFC7231, Section 6.3.3]

pub const NON_AUTHORITATIVE_INFORMATION: StatusCode

203 Non-Authoritative Information [RFC7231, Section 6.3.4]

pub const NO_CONTENT: StatusCode

204 No Content [RFC7231, Section 6.3.5]

pub const RESET_CONTENT: StatusCode

205 Reset Content [RFC7231, Section 6.3.6]

pub const PARTIAL_CONTENT: StatusCode

206 Partial Content [RFC7233, Section 4.1]

pub const MULTI_STATUS: StatusCode

207 Multi-Status [RFC4918]

pub const ALREADY_REPORTED: StatusCode

208 Already Reported [RFC5842]

pub const IM_USED: StatusCode

226 IM Used [RFC3229]

pub const MULTIPLE_CHOICES: StatusCode

300 Multiple Choices [RFC7231, Section 6.4.1]

pub const MOVED_PERMANENTLY: StatusCode

301 Moved Permanently [RFC7231, Section 6.4.2]

pub const FOUND: StatusCode

302 Found [RFC7231, Section 6.4.3]

pub const SEE_OTHER: StatusCode

303 See Other [RFC7231, Section 6.4.4]

pub const NOT_MODIFIED: StatusCode

304 Not Modified [RFC7232, Section 4.1]

pub const USE_PROXY: StatusCode

305 Use Proxy [RFC7231, Section 6.4.5]

pub const TEMPORARY_REDIRECT: StatusCode

307 Temporary Redirect [RFC7231, Section 6.4.7]

pub const PERMANENT_REDIRECT: StatusCode

308 Permanent Redirect [RFC7238]

pub const BAD_REQUEST: StatusCode

400 Bad Request [RFC7231, Section 6.5.1]

pub const UNAUTHORIZED: StatusCode

401 Unauthorized [RFC7235, Section 3.1]

pub const PAYMENT_REQUIRED: StatusCode

402 Payment Required [RFC7231, Section 6.5.2]

pub const FORBIDDEN: StatusCode

403 Forbidden [RFC7231, Section 6.5.3]

pub const NOT_FOUND: StatusCode

404 Not Found [RFC7231, Section 6.5.4]

pub const METHOD_NOT_ALLOWED: StatusCode

405 Method Not Allowed [RFC7231, Section 6.5.5]

pub const NOT_ACCEPTABLE: StatusCode

406 Not Acceptable [RFC7231, Section 6.5.6]

pub const PROXY_AUTHENTICATION_REQUIRED: StatusCode

407 Proxy Authentication Required [RFC7235, Section 3.2]

pub const REQUEST_TIMEOUT: StatusCode

408 Request Timeout [RFC7231, Section 6.5.7]

pub const CONFLICT: StatusCode

409 Conflict [RFC7231, Section 6.5.8]

pub const GONE: StatusCode

410 Gone [RFC7231, Section 6.5.9]

pub const LENGTH_REQUIRED: StatusCode

411 Length Required [RFC7231, Section 6.5.10]

pub const PRECONDITION_FAILED: StatusCode

412 Precondition Failed [RFC7232, Section 4.2]

pub const PAYLOAD_TOO_LARGE: StatusCode

413 Payload Too Large [RFC7231, Section 6.5.11]

pub const URI_TOO_LONG: StatusCode

414 URI Too Long [RFC7231, Section 6.5.12]

pub const UNSUPPORTED_MEDIA_TYPE: StatusCode

415 Unsupported Media Type [RFC7231, Section 6.5.13]

pub const RANGE_NOT_SATISFIABLE: StatusCode

416 Range Not Satisfiable [RFC7233, Section 4.4]

pub const EXPECTATION_FAILED: StatusCode

417 Expectation Failed [RFC7231, Section 6.5.14]

pub const IM_A_TEAPOT: StatusCode

418 I'm a teapot [curiously not registered by IANA but RFC2324]

pub const MISDIRECTED_REQUEST: StatusCode

421 Misdirected Request RFC7540, Section 9.1.2

pub const UNPROCESSABLE_ENTITY: StatusCode

422 Unprocessable Entity [RFC4918]

pub const LOCKED: StatusCode

423 Locked [RFC4918]

pub const FAILED_DEPENDENCY: StatusCode

424 Failed Dependency [RFC4918]

pub const UPGRADE_REQUIRED: StatusCode

426 Upgrade Required [RFC7231, Section 6.5.15]

pub const PRECONDITION_REQUIRED: StatusCode

428 Precondition Required [RFC6585]

pub const TOO_MANY_REQUESTS: StatusCode

429 Too Many Requests [RFC6585]

pub const REQUEST_HEADER_FIELDS_TOO_LARGE: StatusCode

431 Request Header Fields Too Large [RFC6585]

pub const UNAVAILABLE_FOR_LEGAL_REASONS: StatusCode

451 Unavailable For Legal Reasons [RFC7725]

pub const INTERNAL_SERVER_ERROR: StatusCode

500 Internal Server Error [RFC7231, Section 6.6.1]

pub const NOT_IMPLEMENTED: StatusCode

501 Not Implemented [RFC7231, Section 6.6.2]

pub const BAD_GATEWAY: StatusCode

502 Bad Gateway [RFC7231, Section 6.6.3]

pub const SERVICE_UNAVAILABLE: StatusCode

503 Service Unavailable [RFC7231, Section 6.6.4]

pub const GATEWAY_TIMEOUT: StatusCode

504 Gateway Timeout [RFC7231, Section 6.6.5]

pub const HTTP_VERSION_NOT_SUPPORTED: StatusCode

505 HTTP Version Not Supported [RFC7231, Section 6.6.6]

pub const VARIANT_ALSO_NEGOTIATES: StatusCode

506 Variant Also Negotiates [RFC2295]

pub const INSUFFICIENT_STORAGE: StatusCode

507 Insufficient Storage [RFC4918]

pub const LOOP_DETECTED: StatusCode

508 Loop Detected [RFC5842]

pub const NOT_EXTENDED: StatusCode

510 Not Extended [RFC2774]

pub const NETWORK_AUTHENTICATION_REQUIRED: StatusCode

511 Network Authentication Required [RFC6585]