Trait http::HttpTryFrom

pub trait HttpTryFrom<T>: Sized + Sealed {
    type Error: Into<Error>;
}

Private trait for the http crate to have generic methods with fallible conversions.

This trait is similar to the TryFrom trait proposed in the standard library, except this is specialized for the http crate and isn’t intended for general consumption.

This trait cannot be implemented types outside of the http crate, and is only intended for use as a generic bound on methods in the http crate.

Required Associated Types

type Error: Into<Error>

Associated error with the conversion this implementation represents.