Struct http::uri::Authority

pub struct Authority { /* fields omitted */ }

Represents the authority component of a URI.

Methods

impl Authority

pub fn from_static(src: &'static str) -> Self

Attempt to convert an Authority from a static string.

This function will not perform any copying, and the string will be checked if it is empty or contains an invalid character.

Panics

This function panics if the argument contains invalid characters or is empty.

Examples

let authority = Authority::from_static("example.com");
assert_eq!(authority.host(), "example.com");

pub fn from_maybe_shared<T>(src: T) -> Result<Self, InvalidUri> where
T: AsRef<[u8]> + 'static,

Attempt to convert a Bytes buffer to a Authority.

This will try to prevent a copy if the type passed is the type used internally, and will copy the data if it is not.

pub fn host(&self) -> &str

Get the host of this Authority.

The host subcomponent of authority is identified by an IP literal encapsulated within square brackets, an IPv4 address in dotted- decimal form, or a registered name. The host subcomponent is case-insensitive.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                        |---------|
                             |
                            host

Examples

let authority: Authority = "example.org:80".parse().unwrap();

assert_eq!(authority.host(), "example.org");

pub fn port(&self) -> Option<Port<&str>>

Get the port part of this Authority.

The port subcomponent of authority is designated by an optional port number following the host and delimited from it by a single colon (":") character. It can be turned into a decimal port number with the as_u16 method or as a str with the as_str method.

abc://username:password@example.com:123/path/data?key=value&key2=value2#fragid1
                                    |-|
                                     |
                                    port

Examples

Authority with port

let authority: Authority = "example.org:80".parse().unwrap();

let port = authority.port().unwrap();
assert_eq!(port.as_u16(), 80);
assert_eq!(port.as_str(), "80");

Authority without port

let authority: Authority = "example.org".parse().unwrap();

assert!(authority.port().is_none());

pub fn port_u16(&self) -> Option<u16>

Get the port of this Authority as a u16.

Example

let authority: Authority = "example.org:80".parse().unwrap();

assert_eq!(authority.port_u16(), Some(80));

pub fn as_str(&self) -> &str

Return a str representation of the authority