Module http::version

HTTP version

This module contains a definition of the Version type. The Version type is intended to be accessed through the root of the crate (http::Version) rather than this module.

The Version type contains constants that represent the various versions of the HTTP protocol.

Examples

use http::Version;

let http11 = Version::HTTP_11;
let http2 = Version::HTTP_2;
assert!(http11 != http2);

println!("{:?}", http2);

Structs

Version
Represents a version of the HTTP spec.